Ensure parsed source always contains a newline
Some languages require the source text to end in a newline. For example,
when parsing markdown the code-blocks will fail to render correctly if
they are placed at the end and with no newline.

This change ensures all source text (including injected regions)
contains a newline before being parsed.

Any injected regions and highlights are internally remapped to exclude
this appended newline (if there was one appended) before being returned
back to the caller.