Enforce newline for highlighting, then trim
Some grammars require the content to end with a newline for it to parse
correctly. Looking at you, Markdown.

This ensures the content we feed into the highlighter contains a single
newline, which is then trimmed off afterwards.

There is still a remaining issue with injected regions both:

1. Not containing a newline (so they suffer from the same problem).
2. Retaining whatever weird offsets/indentation they have from their
   parent language.

Both of these issues need to be solved in the highlighter crate itself.